# services/sql-assessment-service/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SQL Assessment Service API",
    "version": "1.0.0",
    "description": "REST API for SQL exercise generation, student query grading, and natural-language description of SQL queries in an educational context.",
    "contact": {
      "name": "ALADIN Project"
    }
  },
...
  "paths": {
...
    "/api/grading/grade": {
      "post": {
        "summary": "Full orchestrated grading of a student SQL query",
        "description": "Executes both the reference and student queries, compares their result sets, AST structure, and execution plans, calculates a grade, and optionally generates a natural-language description of the student query. Accepts either a single `referenceQuery` (deprecated) or a `referenceQueries` array; when multiple references are supplied the structurally closest one is selected automatically.\n",
        "tags": [
          "Grading"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GradeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Grading completed successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GradeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, unregistered database, or connection failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Grading pipeline failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/grading/compare/result-set": {
      "post": {
        "summary": "Compare result sets of two SQL queries",
        "description": "Executes both the reference and student queries against the registered database and compares their returned rows. Useful as a lightweight correctness check without full grading overhead.\n",
        "tags": [
          "Grading"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComparisonRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result-set comparison completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultSetComparisonResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or connection failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Comparison failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/grading/compare/ast": {
      "post": {
        "summary": "Compare AST structures of two SQL queries",
        "description": "Parses both the reference and student queries and compares their abstract syntax trees at the structural level (SELECT columns, alias resolution, unsupported structure detection). No database execution is required.\n",
        "tags": [
          "Grading"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComparisonRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AST comparison completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ASTComparisonResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or parse failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Comparison failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/grading/compare/execution-plan": {
      "post": {
        "summary": "Compare query execution plans",
        "description": "Runs EXPLAIN (FORMAT JSON) for both queries and diffs the resulting plans element by element (WHERE, GROUP BY, ORDER BY, JOIN). Returns per-element feedback and a penalty point total.\n",
        "tags": [
          "Grading"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComparisonRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Execution-plan comparison completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionPlanComparisonResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or parse failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Comparison failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "Human-readable error description (localised).",
            "example": "Unable to connect to the database."
          },
          "code": {
            "type": "string",
            "description": "Optional machine-readable error code.",
            "example": "NON_SELECT"
          }
        }
      },
...
      "GradeRequest": {
        "type": "object",
        "required": [
          "connectionInfo",
          "gradingRequest"
        ],
        "properties": {
          "connectionInfo": {
            "$ref": "#/components/schemas/PostgresConnectionInfo"
          },
          "gradingRequest": {
            "type": "object",
            "required": [
              "studentQuery"
            ],
            "properties": {
              "referenceQuery": {
                "type": "string",
                "deprecated": true,
                "description": "Single reference query string. Deprecated — prefer referenceQueries.\n"
              },
              "referenceQueries": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ReferenceQuery"
                },
                "description": "One or more reference solutions. The closest one is selected automatically.\n"
              },
              "studentQuery": {
                "type": "string",
                "description": "The student SQL query to grade."
              }
            }
          },
          "languageCode": {
            "type": "string",
            "example": "en"
          },
          "generationStrategy": {
            "type": "string",
            "enum": [
              "template",
              "llm",
              "hybrid"
            ],
            "description": "Which description generation strategy to use when the student query is not equivalent. Defaults to hybrid / llm depending on query type support.\n"
          },
          "gptOption": {
            "type": "string",
            "enum": [
              "default",
              "creative",
              "multi-step"
            ],
            "description": "GPT option when generationStrategy is llm. Defaults to default."
          }
        }
      },
      "GradeResponse": {
        "type": "object",
        "required": [
          "comparisonResult"
        ],
        "properties": {
          "comparisonResult": {
            "type": "object",
            "properties": {
              "grade": {
                "type": "number",
                "description": "Numeric grade awarded (0–100).",
                "example": 75
              },
              "equivalent": {
                "type": "boolean",
                "description": "Whether the student query is semantically equivalent to the reference."
              },
              "supportedQueryType": {
                "type": "boolean",
                "description": "Whether the query type is supported for structural comparison."
              },
              "feedbackDetails": {
                "type": "object",
                "description": "Detailed per-dimension feedback assembled by the feedback assembler."
              }
            }
          }
        }
      },
      "ComparisonRequest": {
        "type": "object",
        "required": [
          "connectionInfo",
          "studentQuery"
        ],
        "properties": {
          "connectionInfo": {
            "$ref": "#/components/schemas/PostgresConnectionInfo"
          },
          "referenceQuery": {
            "type": "string",
            "deprecated": true,
            "description": "Single reference query. Deprecated — prefer referenceQueries."
          },
          "referenceQueries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReferenceQuery"
            }
          },
          "studentQuery": {
            "type": "string"
          },
          "languageCode": {
            "type": "string",
            "example": "en"
          }
        }
      },
      "ResultSetComparisonResponse": {
        "type": "object",
        "required": [
          "match"
        ],
        "properties": {
          "match": {
            "type": "boolean",
            "description": "Whether both queries return identical result sets."
          },
          "feedback": {
            "type": "object",
            "description": "Optional verdict feedback when the result sets differ."
          }
        }
      },
      "ASTComparisonResponse": {
        "type": "object",
        "required": [
          "columnsMatch",
          "supported"
        ],
        "properties": {
          "columnsMatch": {
            "type": "boolean",
            "description": "Whether the SELECT column lists of both queries match."
          },
          "supported": {
            "type": "boolean",
            "description": "Whether the query uses a supported structure. When false, only result-set equivalence is meaningful for grading.\n"
          },
          "feedback": {
            "type": "object",
            "description": "Per-element AST feedback."
          }
        }
      },
...
    }
  }
}